+++ examples/src/virtual/react/list/prepend-items/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/virtual-examples-react-list-prepend-items
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @af-utils/virtual-core@0.0.22
+  - @af-utils/virtual-react@0.0.24
+
 ## 0.0.3
 
 ### Patch Changes