+++ packages/virtual/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @af-utils/react-virtual-list
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @af-utils/virtual-core@0.0.22
+
 ## 0.0.23
 
 ### Patch Changes